test/ar/unrn/eje1/ClienteTest.java
package ar.unrn.eje1;

import static org.junit.jupiter.api.Assertions.assertEquals;

import org.junit.jupiter.api.Test;

class ClienteTest {

	@Test
	void testCalcularDeudaYPuntosObtenidos() {

		Libro elTunel = new Libro("El Tunel", Libro.REGULARES);
		Libro antesDelFin = new Libro("Antes del Fin", Libro.REGULARES);
		CopiaLibro elTunelCopia = new CopiaLibro(elTunel);
		CopiaLibro antesDelFinCopia = new CopiaLibro(antesDelFin);
		Alquiler alquilerElTunel = new Alquiler(elTunelCopia, 5);
		Alquiler alquilerAntesDelFin = new Alquiler(antesDelFinCopia, 3);
		Cliente yo = new Cliente("Javier");
		yo.alquilar(alquilerElTunel);
		yo.alquilar(alquilerAntesDelFin);
		Object[] resultado = yo.calcularDeudaYPuntosObtenidos();
		System.out.println(resultado[0]);
		System.out.println(resultado[1]);

		assertEquals(10.0, resultado[0]);
		assertEquals(2, resultado[1]);
	}

}
